fix sign compare warning

+++ uhd_wrapper/lib/usrp.cpp
@@ -169,8 +169,7 @@ void Usrp::setRfConfig(const RfConfig &conf) {
 void Usrp::configureRxStreamer(const RfConfig &conf) {
     uhd::stream_args_t rxStreamArgs("fc32", "sc16");
     rxStreamArgs.channels = std::vector<size_t>({});
-    for (size_t rxAntennaIdx = 0; rxAntennaIdx < conf.noRxAntennas;
-         rxAntennaIdx++)
+    for (int rxAntennaIdx = 0; rxAntennaIdx < conf.noRxAntennas; rxAntennaIdx++)
         rxStreamArgs.channels.push_back(rxAntennaIdx);
     rxStreamer_ = usrpDevice_->get_rx_stream(rxStreamArgs);
 }
